+++ platform/src/test/java/org/bytedeco/javacv/FrameGrabberTest.java
@@ -126,6 +126,7 @@ public class FrameGrabberTest {
 
     @Test
     public void testFFmpegFrameGrabberLockingTest() {
+        final boolean[] failed = {false};
         final int numberOfInstances = 20;
         System.out.println("FFmpegFrameGrabberLocking");
 
@@ -139,7 +140,7 @@ public class FrameGrabberTest {
 
                     File tempFile = new File(Loader.getTempDir(), "test" + instance_final + ".mkv");
                     try {
-                        FFmpegFrameRecorder recorder = new FFmpegFrameRecorder(tempFile, 640, 480, 2);
+                        FFmpegFrameRecorder recorder = new FFmpegFrameRecorder(new FileOutputStream(tempFile), 640, 480, 2);
                         recorder.setFormat("matroska"); // mp4 doesn't support streaming
                         recorder.setPixelFormat(AV_PIX_FMT_BGR24);
                         recorder.setVideoCodecName("libx264rgb");
@@ -177,7 +178,7 @@ public class FrameGrabberTest {
 
                         Thread.sleep(1000);
 
-                        FFmpegFrameGrabber grabber = new FFmpegFrameGrabber(tempFile);
+                        FFmpegFrameGrabber grabber = new FFmpegFrameGrabber(new FileInputStream(tempFile));
                         grabber.setSampleMode(FrameGrabber.SampleMode.FLOAT);
                         grabber.start();
 
@@ -213,7 +214,8 @@ public class FrameGrabberTest {
                         grabber.restart();
                         grabber.stop();
                         grabber.release();
-                    } catch (Exception e) {
+                    } catch (Error | Exception e) {
+                        failed[0] = true;
                         fail("Exception should not have been thrown: " + e);
                     } finally {
                         tempFile.delete();
@@ -255,5 +257,6 @@ public class FrameGrabberTest {
                 break;
             }
         }
+        assertFalse(failed[0]);
     }
 }